requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
cairocffi @ file:///C:/Users/evlac/pipwin/cairocffi-1.3.0-cp310-cp310-win_amd64.whl
CairoSVG==2.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cssselect2==0.7.0
defusedxml==0.7.1
docopt==0.6.2
frozenlist==1.3.3
idna==3.4
Js2Py==0.74
magic-filter==1.0.9
multidict==6.0.2
numpy==1.23.5
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pipwin==0.5.2
pycairo==1.22.0
pycparser==2.21
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyvips==2.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tinycss2==1.2.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
webencodings==0.5.1
yarl==1.8.1
